4/taller-ci-cd/api-inferencia/app/pyproject.toml
[project]
name = "inference-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.6",
    "scikit-learn==1.6.1",
    "pandas==2.1.4",
    "prometheus-client==0.21.1",
    "pyyaml"
]